% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_harvest.R
\name{anim_prod_sales}
\alias{anim_prod_sales}
\title{Produces table for sales and production of milk or eggs}
\usage{
anim_prod_sales(
  livestock_df,
  cases,
  anim_id_var,
  anim_vals,
  produced_var,
  produced_val = 1,
  sold_var,
  sold_val = 1,
  amt_sold_vars,
  amt_sold_dk_val,
  group_var,
  table_title,
  anim_type_label
)
}
\arguments{
\item{livestock_df}{Data frame. Livestock ownership and egg and milk production and sales.}

\item{cases}{Data frame. Case to include in analysis}

\item{anim_id_var}{Atomic character vector. Name of ID variable for livestock roster.}

\item{anim_vals}{Numeric vector. Values of \code{anim_id_var} that identify milk- or egg-producing animals, respectively.}

\item{produced_var}{Atomic character vector. Variable for egg/milk production.}

\item{produced_val}{Atomic numeric vector. Value to \code{produced_var} to indicate production.}

\item{sold_var}{Atomic character vector. Variable for egg/milk sales.}

\item{sold_val}{Atomic numeric vector. Value to \code{sold_var} to indicate sales.}

\item{amt_sold_vars}{Character vector. Name variable(s) that capture sales revenue.
In some cases, a single variable; in others, one variable for total value, another for unit value.}

\item{amt_sold_dk_val}{Atomic numeric vector. Value of "do not know" (DK) option for sales revenue.}

\item{group_var}{Atomic character vector. Name of the grouping variable}

\item{table_title}{Atomic character vector. Title of the table.}

\item{anim_type_label}{Atomic character vector. Column label of count of milk/egg producing animals.}
}
\value{
{gt} table object
}
\description{
Produces table for sales and production of milk or eggs
}
