% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_harvest.R
\name{cow_displacement}
\alias{cow_displacement}
\title{Create cow displacement table}
\usage{
cow_displacement(
  hhold_df,
  cases,
  animal_var,
  bull_val = 10,
  cow_val = 12,
  steer_heifer_val = 13,
  calf_val = 14,
  group_var
)
}
\arguments{
\item{hhold_df}{Data frame. All household-level variables.}

\item{cases}{Data frame. Case to include in analysis}

\item{animal_var}{Atomic character vector.
Name of the mulit-select livestock ownership variable,
as it appears in Designer.}

\item{bull_val}{Atomic numeric vector. Value of bull answer option.}

\item{cow_val}{Atomic numeric vector. Value of cow answer option.}

\item{steer_heifer_val}{Atomic numeric vector. Value of steer/heifer answer option.}

\item{calf_val}{Atomic numeric vector. Value of calf answer option.}

\item{group_var}{Atomic character vector. Name of the grouping variable}
}
\value{
{gt} table object
}
\description{
Create cow displacement table
}
