% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{create_col_labels}
\alias{create_col_labels}
\title{Create column labels}
\usage{
create_col_labels(labels, values, df, pattern)
}
\arguments{
\item{labels}{Character vector. Value labels.}

\item{values}{Numeric vector. Values corresponding to labels.}

\item{df}{Data frame with target varibles.}

\item{pattern}{Character. Pattern to identify variables to label.}
}
\description{
Create {gt} column labels that combine labels and values in in the following format: \code{"Label (value)"}
}
