% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{create_dummies}
\alias{create_dummies}
\title{Create dummies from a variable}
\usage{
create_dummies(df, var)
}
\arguments{
\item{df}{Data frame}

\item{var}{Bare variable name of a}
}
\value{
Data frame with dummies added
}
\description{
Create a dummy for each level of a labelled variable
}
