% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_harvest.R
\name{crop_labor}
\alias{crop_labor}
\title{Create crop labor table}
\usage{
crop_labor(
  hhold_df,
  cases,
  grew_crops_var,
  grew_crops_val = 1,
  paid_var,
  free_var,
  members_df,
  member_worked_var,
  member_worked_val = 1,
  group_var
)
}
\arguments{
\item{hhold_df}{Data frame. Household-level variables.}

\item{cases}{Data frame. Case to include in analysis.}

\item{grew_crops_var}{Atomic character vector. Name of variable that indicates whether not grew crops.}

\item{grew_crops_val}{Atomic numeric vector. Value of \code{grew_crops_var} to indicate crops are grown.}

\item{paid_var}{Atomic character vector. Name of variable for categories of paid crop labor, as it appears in Designer.}

\item{free_var}{Atomic character vector. Name of variable for categories of free/exchange crop labor, as it appears in Designer.}

\item{members_df}{Data frame. Household member-level data.}

\item{member_worked_var}{Atomic character vector. Name of variable that indicates whether member worked to grow crops.}

\item{member_worked_val}{Atomic numeric vector. Value of \code{member_worked_var} indicating that the member worked.}

\item{group_var}{Atomic character vector. Name of the grouping variable}
}
\description{
Create crop labor table
}
