% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_planting.R
\name{crop_types}
\alias{crop_types}
\title{Make crop types tables}
\usage{
crop_types(
  crop_df,
  cases,
  crop_type_var,
  temp_crop_val = 1,
  perm_crop_val = 2,
  group_var
)
}
\arguments{
\item{crop_df}{Data frame of parcel-plot-crops}

\item{cases}{Data frame of cases to include in analysis.}

\item{crop_type_var}{Charcter. Name of crop type column.}

\item{temp_crop_val}{Numeric. Value of temporary crop option.}

\item{perm_crop_val}{Numeric. Value of permanent crop option.}

\item{group_var}{Character. Name of grouping variable column (e.g., team, region, etc.).}
}
\value{
{gt} table object
}
\description{
Make crop types tables
}
