% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{extract_var_labels}
\alias{extract_var_labels}
\title{Extract labels from value labels}
\usage{
extract_var_labels(df, var)
}
\arguments{
\item{df}{Data frame.}

\item{var}{Bare variable name.}
}
\value{
Numeric vector. Values of a labelled variable.

Character vector. Character labels of a labelled variable.
}
\description{
First, extract the 'labels' attribute from the column of interest.
Then, returns a numeric vector of possible values
Next, keep the names of each vector element--that is, the string value label
}
