% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{extract_var_values}
\alias{extract_var_values}
\title{Extract variable values from value labels}
\usage{
extract_var_values(df, var)
}
\arguments{
\item{df}{Data frame}

\item{var}{Variable of interest}
}
\value{
Numeric vector. Values of a labelled variable.
}
\description{
First, extract the 'labels' attribute from the column of interest.
Then, returns a numeric vector of possible values
Next, remove names of each vector element
}
