% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_harvest.R
\name{income_sources}
\alias{income_sources}
\title{Create table of income sources}
\usage{
income_sources(
  hhold_df,
  crop_var,
  livestock_var,
  temp_crop_df,
  temp_crop_var,
  temp_crop_val = 1,
  perm_crop_df,
  perm_crop_var,
  perm_crop_val = 1,
  processed_df,
  processed_var,
  processed_val = 1,
  anim_df,
  sold_live_anim_var,
  slaughter_anim_var,
  slaughter_anim_val = 1,
  sold_live_poultry_var,
  slaughter_poultry_var,
  slaughter_poultry_val = 1,
  milk_var,
  milk_val = 1,
  eggs_var,
  eggs_val = 1,
  oth_anim_prod_df,
  oth_anim_var,
  oth_anim_val = 1,
  cases,
  group_var
)
}
\arguments{
\item{hhold_df}{Data frame. Household-level data.}

\item{crop_var}{Character. Indicator variable: grow crops.}

\item{livestock_var}{Character. Indicator variable: raise livestock.}

\item{temp_crop_df}{Data frame. Temporary crop disposition data frame.}

\item{temp_crop_var}{Character. Indicator variable: sell temporary crops.}

\item{temp_crop_val}{Numeric. Value: sell crops.}

\item{perm_crop_df}{Data frame. Permanent crop disposition data frame.}

\item{perm_crop_var}{Character. Indicator variable: sell permanent crops.}

\item{perm_crop_val}{Numeric. Value: sell crops.}

\item{processed_df}{Data frame. Processed crops.}

\item{processed_var}{Character. Indicator variable: sell processed crops.}

\item{processed_val}{Numeric. Value: sell processed crop.}

\item{anim_df}{Data frame. Livestock data frame.}

\item{sold_live_anim_var}{Character. Indicator variable: sold live animals.}

\item{slaughter_anim_var}{Character. Indicator variable: sold slaughtered animal.}

\item{slaughter_anim_val}{Numeric. Value: sold slaughtered animal.}

\item{sold_live_poultry_var}{Character. Indicator variable: sold live poultry.}

\item{slaughter_poultry_var}{Character. Indicator variable: sold slaughtered poultry.}

\item{slaughter_poultry_val}{Numeric. Value: sold slaughtered poultry.}

\item{milk_var}{Character. Indicator variable: sold milk.}

\item{milk_val}{Numeric. Value: sold milk.}

\item{eggs_var}{Character. Indicator variable: sold eggs.}

\item{eggs_val}{Numeric. Value: sold eggs.}

\item{oth_anim_prod_df}{Data frame. Roster of other animal products.}

\item{oth_anim_var}{Character. Indicator variable: sold other animal products.}

\item{oth_anim_val}{Numeric. Value: sold other animal products.}

\item{cases}{Data frame. Cases to include in table analysis.}

\item{group_var}{Character. Name of group variable.}
}
\value{
{gt} table object
}
\description{
Create table of income sources
}
