% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_harvest.R
\name{livestock_labor_tbl}
\alias{livestock_labor_tbl}
\title{Create livestock labor table}
\usage{
livestock_labor_tbl(
  hhold_df,
  cases,
  have_anim_var,
  have_anim_val = 1,
  anim_labor_df,
  anim_labor_id_var,
  anim_labor_var,
  anim_labor_none_val = 0,
  hhold_labor_vals = c(1, 2, 3),
  free_labor_val = 4,
  paid_labor_val = 5,
  group_var
)
}
\arguments{
\item{hhold_df}{Data frame. Household-level data.}

\item{cases}{Data frame. Cases to include in analysis.}

\item{have_anim_var}{Character. Indicator variable: whether raise livestock.}

\item{have_anim_val}{Numeric. Indicator value: raise livestock}

\item{anim_labor_df}{Data frame. Roster of livestock labor.}

\item{anim_labor_id_var}{Character. Labor ID variable in livestock labor roster.}

\item{anim_labor_var}{Character. Number of workers variable.}

\item{anim_labor_none_val}{Numeric. Value indicatoing no workers.}

\item{hhold_labor_vals}{Numeric vector. Codes for household labor categories.}

\item{free_labor_val}{Numeric. Value of free labor.}

\item{paid_labor_val}{Numeric. Value of paid labor.}

\item{group_var}{Character. Name of group variable.}
}
\value{
{gt} table object
}
\description{
Create livestock labor table
}
