% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_harvest.R
\name{livestock_ownership}
\alias{livestock_ownership}
\title{Create livestock ownership table}
\usage{
livestock_ownership(hhold_df, animal_var, cases, group_var)
}
\arguments{
\item{hhold_df}{Data frame.}

\item{animal_var}{Atomic character vector.
Name of multi-select yes/no livestock ownership variable,
as it appears in Designer.}

\item{cases}{Data frame. Case to include in analysis}

\item{group_var}{Atomic character vector. Name of the grouping variable}
}
\value{
{gt} table object
}
\description{
Create livestock ownership table
}
