% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{make_multi_select_quos}
\alias{make_multi_select_quos}
\title{Construct a quosure for multi-select questions}
\usage{
make_multi_select_quos(varname, vals)
}
\arguments{
\item{varname}{Character. Variable name, as it appears in SuSo Designer, for the multi-select question of interest.}

\item{vals}{Numeric vector. Answer options}
}
\value{
Quosure
}
\description{
Creates a quosure of the form: \code{varname__1 == 1 | varname__7 == 1}.
}
