% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_harvest.R
\name{other_prod_sales}
\alias{other_prod_sales}
\title{Create table on product production and sales}
\usage{
other_prod_sales(
  hhold_df,
  product_df,
  cases,
  practice_var,
  practice_val = 1,
  products_var,
  sold_var,
  sold_val = 1,
  amt_sold_vars,
  amt_sold_dk_val,
  group_var,
  table_title
)
}
\arguments{
\item{hhold_df}{Data frame. Household-level data.}

\item{product_df}{Data frame. Product-level data.}

\item{cases}{Data frame. Case to include in analysis}

\item{practice_var}{Atomic character vector. Name of ID variable for whether practice production.}

\item{practice_val}{Atomic numeric vector. Value of \code{practice_var} to indicate practicing.}

\item{products_var}{Atomic character vector. Name of variable for items produced, as it appears in Designer.}

\item{sold_var}{Atomic character vector. Variable for egg/milk sales.}

\item{sold_val}{Atomic numeric vector. Value to \code{sold_var} to indicate sales.}

\item{amt_sold_vars}{Character vector. Name variable(s) that capture sales revenue. In some cases, a single variable; in others, one variable for total value, another for unit value.}

\item{amt_sold_dk_val}{Atomic numeric vector. Value of "do not know" (DK) option for sales revenue.}

\item{group_var}{Atomic character vector. Name of the grouping variable}

\item{table_title}{Atomic character vector. Title of the table.}
}
\value{
{gt} table object
}
\description{
Meant for fisheries, aquaculture, and forestry products.
}
