% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_planting.R
\name{parcels_per_hhold}
\alias{parcels_per_hhold}
\title{Make parcels per household table}
\usage{
parcels_per_hhold(parcel_df, cases, use_var = NULL, use_vals = NULL, group_var)
}
\arguments{
\item{parcel_df}{Data frame of parcels.}

\item{cases}{Data frame of cases to include in analysis}

\item{use_var}{Character. Parcel use variable}

\item{use_vals}{Numeric vector. Values that correspond to an agricultural use}

\item{group_var}{Character. Grouping variable}
}
\value{
{gt} table object
}
\description{
Make parcels per household table
}
