% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_harvest.R
\name{perm_crop_harvest}
\alias{perm_crop_harvest}
\title{Create permanent crop harvest table}
\usage{
perm_crop_harvest(
  parcel_plot_crop_df,
  cases,
  crop_id_var,
  crop_vals,
  produce_var,
  harvest_var,
  harvest_val = 1,
  group_var
)
}
\arguments{
\item{parcel_plot_crop_df}{Data frame of parcel-plot-crop-level observations.}

\item{cases}{Data frame of cases to include in analysis. Data frame must contain \code{interview__id} and the grouping variable indicated in the \code{group_var} parameter}

\item{crop_id_var}{Atomic character vector. Name of crop ID variable as a character.}

\item{crop_vals}{Numeric vector. Codes of the crops that are temporary crops.}

\item{produce_var}{Atomic character vector. Name of variable indicating whether the crop was produced.}

\item{harvest_var}{Atomic character vector. Name of variable indicating whether the crop was harvested.}

\item{harvest_val}{Atomic numeric vector. Value of \code{harvest_var} that indicates the crop was harvested.}

\item{group_var}{Atomic character vector. Name of the grouping variable}
}
\value{
{gt} table object
}
\description{
Create permanent crop harvest table
}
