% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_harvest.R
\name{perm_crop_sales}
\alias{perm_crop_sales}
\title{Create permanent crop sales table}
\usage{
perm_crop_sales(
  crops_sales_df,
  cases,
  crop_id_var,
  crop_vals,
  sold_var,
  sold_val = 1,
  amt_sold_vars,
  amt_sold_dk_val,
  group_var
)
}
\arguments{
\item{crops_sales_df}{Data frame. Crop sales data.}

\item{cases}{Data frame of cases to include in analysis. Data frame must contain \code{interview__id} and the grouping variable indicated in the \code{group_var} parameter}

\item{crop_id_var}{Atomic character vector. Name of crop ID variable as a character.}

\item{crop_vals}{Numeric vector. Codes that identify permanent crops.}

\item{sold_var}{Atomic character vector. Name of variable that indicates whether crop was sold.}

\item{sold_val}{Atomic numeric vector. Value of \code{sold_var} that indicates the crop was sold.}

\item{amt_sold_vars}{Character vector. Name variable(s) that capture sales revenue. In some cases, a single variable; in others, one variable for total value, another for unit value.}

\item{amt_sold_dk_val}{Atomic numeric vector. Value of "do not know" (DK) option for sales revenue.}

\item{group_var}{Atomic character vector. Name of the grouping variable}
}
\value{
{gt} table object
}
\description{
Create permanent crop sales table
}
