% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_planting.R
\name{plot_use}
\alias{plot_use}
\title{Make plot use table}
\usage{
plot_use(plot_df, cases, plot_use_var, group_var)
}
\arguments{
\item{plot_df}{Data frame of plots.}

\item{cases}{Data frame of cases to include in analysis.}

\item{plot_use_var}{Character. Name of plot use column.}

\item{group_var}{Name of grouping variable column (e.g., team, region, etc.).}
}
\value{
{gt} table object
}
\description{
Make plot use table
}
