% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_planting.R
\name{plots_per_parcel}
\alias{plots_per_parcel}
\title{Make plots per parcel table}
\usage{
plots_per_parcel(parcel_plot_df, cases, parcel_id_var, group_var)
}
\arguments{
\item{parcel_plot_df}{Data frame of parcel-plots.}

\item{cases}{Data frame of cases to include in analysis.}

\item{parcel_id_var}{Character. Name of parcel ID column.}

\item{group_var}{Character. Name of grouping variable column (e.g., team, region, etc.).}
}
\value{
{gt} table object
}
\description{
Make plots per parcel table
}
