% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_harvest.R
\name{process_crop_prod}
\alias{process_crop_prod}
\title{Create stats on crop processing}
\usage{
process_crop_prod(
  hhold_df,
  temp_crops_var,
  perm_crops_var,
  processed_var,
  processed_val = 1,
  products_var,
  cases,
  product_df,
  sold_var,
  sold_val = 1,
  amt_sold_vars,
  amt_sold_dk_val,
  group_var
)
}
\arguments{
\item{hhold_df}{Household data frame}

\item{temp_crops_var}{Character. Indicator variable: any temporary crop harvested.}

\item{perm_crops_var}{Character. Indicator variable: any permanent crop harvested.}

\item{processed_var}{Character. Indicator variable: any crops processed.}

\item{processed_val}{Numeric. Value of indicator: any crops processed.}

\item{products_var}{Character. Name of variable as in Designer: which processed products.}

\item{cases}{Data frame. Case to include in analysis}

\item{product_df}{Data frame. roster of processed crops.}

\item{sold_var}{Character. Indicator variable: whether product sold.}

\item{sold_val}{Numeric. Value of indicator: product sold.}

\item{amt_sold_vars}{Character vector. Name(s) of variables with sales values.}

\item{amt_sold_dk_val}{Numeric. Value of "do not know" value.}

\item{group_var}{Character. Name of group variable.}
}
\value{
{gt} table object
}
\description{
Create stats on crop processing
}
