% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{replace_nan}
\alias{replace_nan}
\title{Replace NaN with NA_real_}
\usage{
replace_nan(df)
}
\arguments{
\item{df}{Data frame}
}
\description{
Replace NaN, which results from division by 0, with NA. This makes data destined for tables easier to handle.
}
