% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_harvest.R
\name{sector_labor}
\alias{sector_labor}
\title{Create table for sector labor}
\usage{
sector_labor(
  hhold_df,
  produce_var,
  produce_val = 1,
  cases,
  labor_var,
  hhold_labor_vals = c(1, 2, 3),
  free_labor_val = 4,
  paid_labor_val = 5,
  group_var,
  table_title,
  sector_label
)
}
\arguments{
\item{hhold_df}{Data frame. Household-level data.}

\item{produce_var}{Character. Indicator variable: whether engaged in sector.}

\item{produce_val}{Numeric. Value: engaged in sector.}

\item{cases}{Data frame. Cases to include in table computations.}

\item{labor_var}{Character. Name of labor categories used. Name of multi-select question as in Designer.}

\item{hhold_labor_vals}{Numeric vector. Codes of answers corresponding to household labor.}

\item{free_labor_val}{Numeric. Code of free labor.}

\item{paid_labor_val}{Numeric. Code of paid labor.}

\item{group_var}{Character. Name of group variable.}

\item{table_title}{Character. Title of table.}

\item{sector_label}{Character. Label of first column that counts households engaged in sector.}
}
\value{
{gt} table object
}
\description{
Create table for sector labor
}
