% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{style_table}
\alias{style_table}
\title{Apply style to tables}
\usage{
style_table(
  df,
  heading_color = "#0F2B1D",
  column_label_color = "#264535",
  row_group_color = "#516A5D"
)
}
\arguments{
\item{df}{Data frame}

\item{heading_color}{Character. Hex color for table header background color.}

\item{column_label_color}{Character. Hex color for column label background color.}

\item{row_group_color}{Character. Hex color for row group background color.}
}
\description{
Apply styles from \code{gt::tab_options()} to tables
}
