% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_harvest.R
\name{hen_displacement}
\alias{hen_displacement}
\title{Create hen displacement table}
\usage{
hen_displacement(
  hhold_df,
  cases,
  animal_var,
  cock_val = 51,
  hen_val = 52,
  pullet_val = 53,
  group_var
)
}
\arguments{
\item{hhold_df}{Data frame. All household-level variables.}

\item{cases}{Data frame. Case to include in analysis}

\item{animal_var}{Atomic character vector. Name of livestock ownership variable, as it appears in Designer.}

\item{cock_val}{Atomic numeric vector. Value of cock/broiler answer option.}

\item{hen_val}{Atomic numeric vector. Value of hen answer option.}

\item{pullet_val}{Atomic numeric vector. Value of pullet answer option.}

\item{group_var}{Atomic character vector. Name of the grouping variable}
}
\value{
{gt} table object
}
\description{
Create hen displacement table
}
